+++ Modules/core/Controller/CoreusersController.php
@@ -131,8 +131,6 @@ class CoreusersController extends CoresecureController
         $form->setValidationButton(CoreTranslator::Save($lang), "coreusersedit/" . $id);
         $form->setCancelButton(CoreTranslator::Cancel($lang), "coreusers");
 
-
-
         $rolesTableHtml  = '';
         if ($id > 0) {
             $formPwd = new Form($this->request, "coreuseretidpwd");
@@ -142,14 +140,12 @@ class CoreusersController extends CoresecureController
             $formPwd->addPassword("pwdconfirm", CoreTranslator::New_password($lang));
             $formPwd->setValidationButton(CoreTranslator::Save($lang), "coreusersedit/" . $id);
 
-
-
             $csm = new CoreSpace();
             $roles = $csm->getUserSpacesRoles(0, $user['id']);
             $pum = new CorePendingAccount();
             $pendings = $pum->getSpaceIdsForPending($user['id']);
             foreach ($pendings as $p) {
-                $roles[] = ['space_name' => $p['space_name'], 'role_name' => CoreTranslator::PendingUserAccount($lang)];
+                $roles[] = ['id_space' => $p['id_space'], 'space_name' => $p['space_name'], 'role_name' => CoreTranslator::PendingUserAccount($lang)];
             }
             $rolesTable = new TableView('spaces');
             $rolesTable->setTitle(CoreTranslator::Spaces($lang));
@@ -157,6 +153,7 @@ class CoreusersController extends CoresecureController
                 "space_name" => CoreTranslator::Space($lang),
                 "role_name" => CoreTranslator::Role($lang),
             );
+            $rolesTable->addLineEditButton('/corespaceuseredit/{{id_space}}/'.$user['id'], 'id_space');
             $rolesTableHtml = $rolesTable->view($roles, $headers);
         }
         $script = "";
@@ -165,29 +162,25 @@ class CoreusersController extends CoresecureController
 
             if (!$form->getParameter("email") || !$modelUser->isEmailFormat($form->getParameter("email"))) {
                 $canEditUser = false;
-                $this->displayFormWarnings("EmailInvalid", $id, $lang);
-                return;
+                return $this->displayFormWarnings("EmailInvalid", $id, $lang);
             }
             if (!$id) {
                 // creating a new user
                 if ($modelUser->isLogin($this->request->getParameter('login'))) {
                     $canEditUser = false;
-                    $this->displayFormWarnings("LoginAlreadyExists", $id, $lang);
-                    return;
+                    return $this->displayFormWarnings("LoginAlreadyExists", $id, $lang);
                 }
                 if ($modelUser->isEmail($form->getParameter("email"))) {
                     // if email already exists, warn user
                     $canEditUser = false;
-                    $this->displayFormWarnings("EmailAlreadyExists", $id, $lang);
-                    return;
+                    return $this->displayFormWarnings("EmailAlreadyExists", $id, $lang);
                 }
             } else {
                 // updating an existing user
                 if ($modelUser->isEmail($form->getParameter("email")) && ($form->getParameter("email") != $user["email"])) {
                     // if email, excepting user's one, already exists, warn user
                     $canEditUser = false;
-                    $this->displayFormWarnings("EmailAlreadyExists", $id, $lang);
-                    return;
+                    return $this->displayFormWarnings("EmailAlreadyExists", $id, $lang);
                 }
             }
 
@@ -204,8 +197,7 @@ class CoreusersController extends CoresecureController
 
         if ($id > 0 && $formPwd->check()) {
             $this->editPwdQuery($form, $lang);
-            $this->redirect("coreusers");
-            return;
+            return $this->redirect("coreusers");
         }
 
         $formPwdHtml = "";
@@ -226,7 +218,7 @@ class CoreusersController extends CoresecureController
     {
         $_SESSION["flash"] = CoreTranslator::$cause($lang);
         $_SESSION["flashClass"] = "danger";
-        $this->redirect("coreusersedit/" . $id ?? 0);
+        return $this->redirect("coreusersedit/" . $id ?? 0);
     }
 
     protected function editPwdQuery($formPwd, $lang)
@@ -417,5 +409,32 @@ class CoreusersController extends CoresecureController
             'lang' => $lang,
             'form' => $form->getHtml($lang)
         ));
+    }
+
+
+    public function impersonateAction($id_user)
+    {
+        $this->checkAuthorization(CoreStatus::$ADMIN);
+
+        $this->request->getSession()->setAttribut("logged_id_user", $_SESSION['id_user']);
+        $this->request->getSession()->setAttribut("logged_login", $_SESSION['login']);
+        $this->request->getSession()->setAttribut("logged_email", $_SESSION['email']);
+        $this->request->getSession()->setAttribut("logged_user_status", $_SESSION['user_status']);
+
+        $modelUser = new CoreUser();
+        $user = $modelUser->getInfo($id_user);
+
+        Configuration::getLogger()->debug('[impersonate][admin]', [
+            'to_id' => $user['id'], 'to_login' => $user['login'],
+            'from_id' => $_SESSION['id_user'], 'from_login' => $_SESSION['login']
+        ]);
+
+
+        $this->request->getSession()->setAttribut("id_user", $user['id']);
+        $this->request->getSession()->setAttribut("login", $user['login']);
+        $this->request->getSession()->setAttribut("email", $user['email']);
+        $this->request->getSession()->setAttribut("user_status", CoreStatus::$USER);
+
+        $this->redirect("coretiles");
     }
 }