+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changes
 
+## 2.9.7
+
+### Features
+
+* [core] coreusersedit, add link to edit user in related space
+* [core] as admin, allow full user impersonification in coreusersedit
+
+### Fixes
+
+* [core] differentiate pdo from other errors in db handling
+* [core] remove request access button for private spaces in search results
+
 ## 2.9.6
 
 ### Fixes